#179 #180 [-] Replaced all notions of vnlist / votelist / wishlist in the app with the new ulist -> changed database schema, repositories and API requests accordingly. [-] #178 : Removed "My votelist" and "My wishlist" pages. [-] Removed "High" / "Medium" / "Low" wishlist priorities in favor of a single "Wishlist" option. [#] Changed database schema to account for the new ulist object. [+] VN details bottom sheet is now able to show multiple selected labels within the same group, and easily select a single or multiple labels within a group.

+++ app/src/main/java/com/booboot/vndbandroid/ui/hometabs/HomeTabsAdapter.kt
@@ -16,7 +16,7 @@ class HomeTabsAdapter(fm: FragmentManager, private val type: Int) : BaseFragment
 
     override fun getItem(position: Int) = (fragments[position] ?: VNListFragment()).apply {
         arguments = Bundle().apply {
-            putInt(TAB_VALUE_ARG, tabs[position].value)
+            putLong(TAB_VALUE_ARG, tabs[position].value)
             putInt(HomeTabsFragment.LIST_TYPE_ARG, type)
         }
     }
